Small refactor to strip out callbacks
